clean up controllers and templates
