--- hw15-messageSystem/src/main/java/ru/otus/web/controllers/LoginUser.java
@@ -1,14 +1,0 @@
-package ru.otus.web.controllers;
-
-import lombok.Getter;
-import lombok.Setter;
-
-class LoginUser {
-    @Getter
-    @Setter
-    private String login;
-
-    @Getter
-    @Setter
-    private String password;
-}
