+++ hw15-messageSystem/src/main/java/ru/otus/web/messagesystem/MsClient.java
