+++ hw15-messageSystem/src/main/java/ru/otus/web/services/TemplateProcessor.java
@@ -1,6 +1,5 @@
 package ru.otus.web.services;
 
-import java.io.IOException;
 import java.util.Map;
 
 public interface TemplateProcessor {
