--- hw15-messageSystem/src/main/resources/templates/login.html
@@ -1,26 +1,0 @@
-﻿<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<html xmlns="http://www.w3.org/1999/xhtml" content="text/html;">
-<head>
-    <title>Login</title>
-</head>
-<body>
-<div class="top">
-    <img src="img/login.png">
-    <br/>
-    <form th:action="@{/login}" method="post" th:object="${loginUser}">
-        <label>
-            login <br/>
-            <input type="text" name="login"  th:field="*{login}">
-        </label>
-        <br/>
-        <label>
-            password<br/>
-            <input type="password" name="password"  th:field="*{password}">
-        </label>
-        <br/><br/>
-        <input type="submit" value="Перейти к списку пользователей">
-    </form>
-</div>
-</body>
-</html>
